drm: samsung: compression source is moved to modifier.
The interface of compression source is changed from plane property
to modifier because this information is related to buffer.

This information is useful if there is a problem with AFBC
compressed buffer.

+++ drivers/gpu/drm/samsung/exynos_drm_decon.h
@@ -83,7 +83,7 @@ struct dpu_bts_win_config {
 	bool is_afbc;
 	int dpp_ch;
 	u32 format;
-	u16 comp_src;
+	u64 comp_src;
 };
 
 struct dpu_bts {
@@ -157,7 +157,7 @@ struct dpu_log_dsim_cmd {
 
 struct dpu_log_dpp {
 	u32 id;
-	u16 comp_src;
+	u64 comp_src;
 	u32 recovery_cnt;
 };
